```更改CategoryRepository中的主键类型并引入分页响应类将CategoryRepository中的主键类型从Integer更改为Long，以适应后端架构更改。 引入新的Page类，用于在需要时支持分页处理，为前端提供更灵活的数据检索方式。 ```

+++ src/main/java/com/ovo307000/bigevent/repository/user/CategoryRepository.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Repository("userCategoryRepository")
-public interface CategoryRepository extends JpaRepository<CategoryDTO, Integer>
+public interface CategoryRepository extends JpaRepository<CategoryDTO, Long>
 {
     List<CategoryDTO> findCategoryDTOByCreateUser(UserDTO userDTO);
